% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_DiscSurv.R
\name{coef.DiscSurv}
\alias{coef.DiscSurv}
\title{Extract coefficients of a DiscSurv object}
\usage{
\method{coef}{DiscSurv}(object, lambda, which = 1:length(object$lambda), drop = TRUE, ...)
}
\arguments{
\item{object}{a \code{DiscSurv} object.}

\item{lambda}{values of the regularization parameter lambda at which coefficients are requested. For values of lambda not in the sequence of fitted models, linear interpolation is used.}

\item{which}{indices of the penalty parameter lambda at which predictions are required. By default, all indices are returned. If lambda is specified, this will override which.}

\item{drop}{whether to keep coefficient names}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Return the model coefficients of a \code{DiscSurv} object
}
\examples{
data(DiscTime)
data <- DiscTime$data
Event.char <- DiscTime$Event.char
Z.char <- DiscTime$Z.char
Time.char <- DiscTime$Time.char
fit <- DiscSurv(data, Event.char, Z.char, Time.char)
coef(fit, lambda = fit$lambda)$alpha[, 1:10]
coef(fit, lambda = fit$lambda)$beta[, 1:10]
}
