% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_pplasso.R
\name{coef.ppLasso}
\alias{coef.ppLasso}
\alias{coef.gr_ppLasso}
\title{Extract coefficients of a ppLasso or gr_ppLasso object}
\usage{
\method{coef}{ppLasso}(object, lambda, which = 1:length(object$lambda), drop = TRUE, ...)

\method{coef}{gr_ppLasso}(object, lambda, which = 1:length(object$lambda), drop = TRUE, ...)
}
\arguments{
\item{object}{a \code{ppLasso} or \code{gr_ppLasso} object.}

\item{lambda}{values of the regularization parameter lambda at which coefficients are requested. For values of lambda not in the sequence of fitted models, linear interpolation is used.}

\item{which}{indices of the penalty parameter lambda at which predictions are required. By default, all indices are returned. If lambda is specified, this will override which.}

\item{drop}{whether to keep coefficient names}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Return the model coefficients of a \code{ppLasso} or \code{gr_ppLasso} object
}
\examples{
#fit glm without grouped covariates
data(BinaryData)
data <- BinaryData$data
Y.char <- BinaryData$Y.char
prov.char <- BinaryData$prov.char
Z.char <- BinaryData$Z.char
fit <- pp.lasso(data, Y.char, Z.char, prov.char)
coef(fit, lambda = fit$lambda)$beta[, 1:10]
coef(fit, lambda = fit$lambda)$gamma[1:10, 1:5]
#fit glm with grouped covariates
data(BinaryData)
data <- BinaryData$data
Y.char <- BinaryData$Y.char
prov.char <- BinaryData$prov.char
Z.char <- BinaryData$Z.char
group <- BinaryData$group
fit <- grp.lasso(data, Y.char, Z.char, prov.char, group = group)
coef(fit, lambda = fit$lambda)$beta[, 1:5]
coef(fit, lambda = fit$lambda)$gamma[1:10, 1:5]
}
