% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.grLasso.R
\name{cv.grp.lasso}
\alias{cv.grp.lasso}
\title{cross-validation for grp.lasso}
\usage{
cv.grp.lasso(
  data,
  Y.char,
  Z.char,
  prov.char,
  group = 1:length(Z.char),
  ...,
  nfolds = 10,
  seed,
  fold,
  trace.cv = FALSE
)
}
\arguments{
\item{data}{an `dataframe` or `list` object that contains the variables in the model.}

\item{Y.char}{name of the response variable from `data` as a character string, as in `pp.lasso` function.}

\item{Z.char}{names of covariates from `data` as vector of character strings, as in `pp.lasso` function.}

\item{prov.char}{name of provider IDs variable from `data` as a character string, as in `pp.lasso` function.}

\item{group}{a vector describing the grouping of the coefficients. If there are coefficients to be included in the model without being penalized, assign them to group 0 (or "0").}

\item{...}{extra arguments to be passed to function.}

\item{nfolds}{the number of cross-validation folds. Default is 10.}

\item{seed}{the seed of the random number generator in order to obtain reproducible results.}

\item{fold}{a vector that specifies the fold that observations belongs to. By default the observations are randomly assigned.}

\item{trace.cv}{`cv.pp.lasso` will provide user with the progress of cross validation if `trace.cv = TRUE`. Default is FALSE.}
}
\value{
An object with S3 class `cv.gr_ppLasso`.

\item{cve}{the error for each value of lambda, averaged across the cross-validation folds.}

\item{cvse}{the estimated standard error associated with each value of for cve.}

\item{lambda}{the sequence of regularization parameter values along which the cross-validation error was calculated.}

\item{fit}{the fitted `pp.lasso` object for the whole data.}

\item{fold}{the fold assignments for cross-validation for each observation}

\item{min}{the index of lambda corresponding to lambda.min.}

\item{lambda.min}{the value of lambda with the minimum cross-validation error.}
}
\description{
performs k-fold cross validation for group penalized regression models over a grid of values of regularization parameter lambda.
}
\examples{
data(GLM_Data)
Y.char <- 'Y'
prov.char <- 'Prov.ID'
Z.char <- c("Z1", "Z2", "Z3", "Z4", "Z5")
group <- c(1, 1, 2, 2, 3)
fit <- cv.grp.lasso(GLM_Data, Y.char, Z.char, prov.char, group = group)
# the best lambda using cross validation
fit$lambda.min

}
\references{
K. He, J. Kalbfleisch, Y. Li, and et al. (2013) Evaluating hospital readmission rates in dialysis facilities; adjusting for hospital effects.
\emph{Lifetime Data Analysis}, \strong{19}: 490-512.
\cr
}
