% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.ppLasso.R
\name{cv.pp.lasso}
\alias{cv.pp.lasso}
\title{Cross-validation for penalized generalized regression model}
\usage{
cv.pp.lasso(
  data,
  Y.char,
  Z.char,
  prov.char,
  penalize.x = rep(1, length(Z.char)),
  ...,
  nfolds = 10,
  seed,
  fold,
  trace.cv = FALSE
)
}
\arguments{
\item{data}{an \code{dataframe} or \code{list} object that contains the variables in the model.}

\item{Y.char}{name of the response variable from \code{data} as a character string, as in \code{pp.lasso} function.}

\item{Z.char}{names of covariates from \code{data} as vector of character strings, as in \code{pp.lasso} function.}

\item{prov.char}{name of provider IDs variable from \code{data} as a character string, as in \code{pp.lasso} function.}

\item{penalize.x}{a vector indicates whether the corresponding covariate will be penalized, as in \code{pp.lasso} function.}

\item{...}{extra arguments to be passed to function.}

\item{nfolds}{the number of cross-validation folds. Default is 10.}

\item{seed}{the seed of the random number generator in order to obtain reproducible results.}

\item{fold}{a vector that specifies the fold that observations belongs to. By default the observations are randomly assigned.}

\item{trace.cv}{\code{cv.pp.lasso} will provide user with the progress of cross validation if \code{trace.cv = TRUE}. Default is FALSE.}
}
\value{
An object with S3 class \code{cv.ppLasso}.

\item{cve}{the error for each value of lambda, averaged across the cross-validation folds.}

\item{cvse}{the estimated standard error associated with each value of for cve.}

\item{lambda}{the sequence of regularization parameter values along which the cross-validation error was calculated.}

\item{fit}{the fitted \code{pp.lasso} object for the whole data.}

\item{fold}{the fold assignments for cross-validation for each observation}

\item{min}{the index of lambda corresponding to lambda.min.}

\item{lambda.min}{the value of lambda with the minimum cross-validation error.}
}
\description{
Performs k-fold cross validation for penalized regression models over a grid of values of regularization parameter lambda.
}
\examples{
data(BinaryData)
data <- BinaryData$data
Y.char <- BinaryData$Y.char
prov.char <- BinaryData$prov.char
Z.char <- BinaryData$Z.char
cv.fit <- cv.pp.lasso(data, Y.char, Z.char, prov.char, nfolds = 10)
# the best lambda using cross validation
cv.fit$lambda.min

}
\references{
K. He, J. Kalbfleisch, Y. Li, and et al. (2013) Evaluating hospital readmission rates in dialysis facilities; adjusting for hospital effects.
\emph{Lifetime Data Analysis}, \strong{19}: 490-512.
\cr
}
