% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.strat_cox.R
\name{cv.strat_cox}
\alias{cv.strat_cox}
\title{Cross-validation for penalized stratified cox model}
\usage{
cv.strat_cox(
  data,
  Event.char,
  Z.char,
  Time.char,
  prov.char,
  group = 1:length(Z.char),
  se = c("quick", "bootstrap"),
  ...,
  nfolds = 10,
  seed,
  fold,
  trace.cv = FALSE
)
}
\arguments{
\item{data}{an \code{dataframe} or \code{list} object that contains the variables in the model.}

\item{Event.char}{name of the event indicator in \code{data} as a character string. Event indicator should be a
binary variable with 1 indicating that the event has occurred and 0 indicating (right) censoring.}

\item{Z.char}{names of covariates in \code{data} as vector of character strings.}

\item{Time.char}{name of the follow up time in \code{data} as a character string.}

\item{prov.char}{name of provider IDs variable in \code{data} as a character string. (can be seen as "stratum")}

\item{group}{a vector describing the grouping of the coefficients. If there are coefficients to be included in the model without being penalized, assign them to group 0 (or "0").}

\item{se}{the method used to compute the standard error of the cross-validation error, either \code{"quick"} or \code{"bootstrap"}.}

\item{...}{extra arguments to be passed to function.}

\item{nfolds}{the number of cross-validation folds. Default is 10.}

\item{seed}{the seed of the random number generator in order to obtain reproducible results.}

\item{fold}{a vector that specifies the fold that observations belongs to. By default the observations are randomly assigned.}

\item{trace.cv}{\code{cv.strat_cox} will provide user with the progress of cross validation if \code{trace.cv = TRUE}. Default is FALSE.}
}
\value{
An object with S3 class \code{cv.strat_cox}.

\item{cve}{the error for each value of lambda, averaged across the cross-validation folds.}

\item{cvse}{the estimated standard error associated with each value of for cve.}

\item{lambda}{the sequence of regularization parameter values along which the cross-validation error was calculated.}

\item{fit}{the fitted \code{strat_cox} object for the whole data.}

\item{fold}{the fold assignments for cross-validation for each observation}

\item{min}{the index of lambda corresponding to lambda.min.}

\item{lambda.min}{the value of lambda with the minimum cross-validation error.}
}
\description{
Performs k-fold cross validation for a penalized stratified cox model over a grid of values of regularization parameter lambda.
}
\examples{
data(ContTime)
data <- ContTime$data
Event.char <- ContTime$Event.char
prov.char <- ContTime$prov.char
Z.char <- ContTime$Z.char
Time.char <- ContTime$Time.char
cv.fit <- cv.strat_cox(data, Event.char, Z.char, Time.char, prov.char,
                       group = c(1, 2, 2, 3, 3), nfolds = 10, se = "quick")
# the best lambda using cross validation
cv.fit$lambda.min

}
\references{
K. He, J. Kalbfleisch, Y. Li, and et al. (2013) Evaluating hospital readmission rates in dialysis facilities; adjusting for hospital effects.
\emph{Lifetime Data Analysis}, \strong{19}: 490-512.
\cr
}
