% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{GLM_Data}
\alias{GLM_Data}
\title{example data for `pp.lasso` and `grp.lasso`}
\format{
A list containing the following elements:
\describe{
  \item{Y}{response variable.}
  \item{Prov.ID}{provider information.}
  \item{Z1, ..., Z5}{5 continuous covariates.}
}
}
\usage{
data(GLM_Data)
}
\description{
A simulated data set containing response variable, provider information and 5 covariates.
}
\keyword{datasets}
