\name{grplasso-package}
\alias{grplasso-package}
\docType{package}
\title{
\packageTitle{grplasso}
}
\description{
\packageDescription{grplasso}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{grplasso}
\packageIndices{grplasso}
The best entry point for the package are the examples in the help
file of the function \code{\link{grplasso}}.
}
\author{
\packageAuthor{grplasso}

Maintainer: \packageMaintainer{grplasso}
}
\references{Lukas Meier, Sara van de Geer and Peter B\"uhlmann (2008), \emph{The
    Group Lasso for Logistic Regression}, Journal of the Royal
  Statistical Society, 70 (1), 53 - 71}
\keyword{ package }

