% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{linear_data}
\alias{linear_data}
\title{Example data for linear model}
\format{
A list containing the following elements:
\describe{
\item{data}{example data.\code{time} represents the observation time; \code{status} is the event indicator; \code{Prov.ID} is the center indicator (include 5 centers); \code{Z1}, ..., \code{Z5} are 5 continuous covariates.}
\item{Y.char}{variable name of the response variable.}
\item{prov.char}{variable name of the provider indicator.}
\item{Z.char}{variable names of covariates.}
}
}
\usage{
data(linear_data)
}
\description{
A simulated data set containing response variable, provider information and 10 covariates.
}
\examples{
data(linear_data)
data <- linear_data$data
head(data)
}
\keyword{datasets}
