% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cve.ppDiscsurv.R
\name{plot.cv.ppDiscSurv}
\alias{plot.cv.ppDiscSurv}
\title{Plot the cross entropy loss from a cv.ppDiscSurv object}
\usage{
\method{plot}{cv.ppDiscSurv}(
  x,
  log.x = TRUE,
  vertical.line = TRUE,
  col.vertical.line = "blue",
  col.dot = "red",
  ...
)
}
\arguments{
\item{x}{a \code{cv.ppDiscSurv} object.}

\item{log.x}{whether the horizontal axis be on the log scale.}

\item{vertical.line}{whether draws a vertical line at the value where cross-validaton error is minimized.}

\item{col.vertical.line}{colour of the vertical line marking the selected lambda.}

\item{col.dot}{colour of the plotted cross-validation error points.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Return the plot of the cross entropy loss from a \code{cv.ppDiscSurv} object
}
\examples{
data(DiscTime)
data <- DiscTime$data
Event.char <- DiscTime$Event.char
prov.char <- DiscTime$prov.char
Z.char <- DiscTime$Z.char
Time.char <- DiscTime$Time.char
cv.fit.ppDiscSurv <- cv.pp.DiscSurv(data, Event.char, prov.char, Z.char, Time.char, nfolds = 10)
plot(cv.fit.ppDiscSurv)
}
