% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.path.Discsurv.R
\name{plot.DiscSurv}
\alias{plot.DiscSurv}
\title{Plot regularization path of coefficients from a DiscSurv object}
\usage{
\method{plot}{DiscSurv}(x, log.x = TRUE, label = FALSE, ...)
}
\arguments{
\item{x}{a \code{DiscSurv} object.}

\item{log.x}{whether the horizontal axis be on the log scale.}

\item{label}{whether annotates the plot with labels.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Return the plot the regularization path from a \code{DiscSurv} object
}
\examples{
data(DiscTime)
data <- DiscTime$data
Event.char <- DiscTime$Event.char
Z.char <- DiscTime$Z.char
Time.char <- DiscTime$Time.char
fit <- DiscSurv(data, Event.char, Z.char, Time.char)
plot(fit, label = TRUE)
}
