% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.path.R
\name{plot.gr_ppLasso}
\alias{plot.gr_ppLasso}
\title{plot regularization path of coefficients from a `gr_ppLasso` object}
\usage{
\method{plot}{gr_ppLasso}(fit, log.x = T, label = F)
}
\arguments{
\item{fit}{a `gr_ppLasso` object.}

\item{log.x}{whether the horizontal axis be on the log scale.}

\item{label}{whether annotates the plot with labels.}
}
\description{
plot the cross entropy Loss from a `gr_ppLasso` object
}
\examples{
data(GLM_Data)
Y.char <- 'Y'
prov.char <- 'Prov.ID'
Z.char <- c("Z1", "Z2", "Z3", "Z4", "Z5")
group <- c(1, 1, 2, 2, 2)
fit <- grp.lasso(GLM_Data, Y.char, Z.char, prov.char, group = group)
plot(fit)
}
