% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ppDiscSurv.R
\name{predict.ppDiscSurv}
\alias{predict.ppDiscSurv}
\title{Predictions of a ppDiscSurv object}
\usage{
\method{predict}{ppDiscSurv}(
  object,
  data,
  Event.char,
  prov.char,
  Z.char,
  Time.char,
  lambda,
  which = 1:length(object$lambda),
  type = c("response", "vars", "nvars"),
  return.Array = TRUE,
  which.lambda = "all",
  ...
)
}
\arguments{
\item{object}{a \code{ppDiscSurv} object.}

\item{data}{an \code{dataframe} or \code{list} object that contains the variables for prediction.}

\item{Event.char}{name of the event indicator in \code{data} as a character string.}

\item{prov.char}{name of provider IDs variable in \code{data} as a character string.}

\item{Z.char}{names of covariates in \code{data} as vector of character strings.}

\item{Time.char}{name of the observation time in \code{data} as a character string.}

\item{lambda}{values of the regularization parameter lambda at which predictions are requested. For values of lambda not in the sequence of fitted models, linear interpolation is used.}

\item{which}{indices of the penalty parameter lambda at which predictions are required. By default, all indices are returned. If lambda is specified, this will override which.}

\item{type}{type of prediction:
\itemize{
\item \code{response}: fitted values (i.e., \code{exp(eta)/(1 + exp(eta))})
\item \code{vars}: the indices for the non-zero coefficients
\item \code{nvars}: the number of non-zero coefficients
}}

\item{return.Array}{whether to return the predictions as a 3-dimensional array indexed by subject, time point and lambda.}

\item{which.lambda}{determine which lambda values are included in the output of the prediction. By default, its value is set to "all," resulting in a matrix of predicted values for each lambda presented as a list. However, if specific numeric values are provided, only the predicted matrix corresponding to those specified values will be included in the output.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Return the model predictions of a \code{ppDiscSurv} object
}
\examples{
data(DiscTime)
data <- DiscTime$data
Event.char <- DiscTime$Event.char
prov.char <- DiscTime$prov.char
Z.char <- DiscTime$Z.char
Time.char <- DiscTime$Time.char
fit <- pp.DiscSurv(data, Event.char, prov.char, Z.char, Time.char)
predict(fit, data, Event.char, prov.char, Z.char, Time.char,
        lambda = fit$lambda, type = "response",
        which.lambda = fit$lambda[1])[1:5, ]
predict(fit, data, Event.char, prov.char, Z.char, Time.char, lambda = 0.04, type = "vars")
}
