% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_strat_cox.R
\name{predict.strat_cox}
\alias{predict.strat_cox}
\title{Predictions of a strat_cox object}
\usage{
\method{predict}{strat_cox}(
  object,
  data,
  Z.char,
  lambda,
  which = 1:length(object$lambda),
  type = c("link", "response", "vars", "nvars", "groups", "ngroups"),
  ...
)
}
\arguments{
\item{object}{a \code{strat_cox} object.}

\item{data}{an \code{dataframe} or \code{list} object that contains the variables for prediction.}

\item{Z.char}{names of covariates in \code{data} as vector of character strings.}

\item{lambda}{values of the regularization parameter lambda at which predictions are requested. For values of lambda not in the sequence of fitted models, linear interpolation is used.}

\item{which}{indices of the penalty parameter lambda at which predictions are required. By default, all indices are returned. If lambda is specified, this will override which.}

\item{type}{type of prediction:
\itemize{
\item \code{link}: linear predictors
\item \code{response}: risk (i.e., \code{exp(link)})
\item \code{vars}: the indices for the non-zero coefficients
\item \code{nvars}: the number of non-zero coefficients
\item \code{groups}: the indices for the non-zero groups
\item \code{ngroups}: the number of non-zero coefficients
}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Return the model predictions of a \code{strat_cox} object
}
\examples{
data(ContTime)
data <- ContTime$data
Event.char <- ContTime$Event.char
prov.char <- ContTime$prov.char
Z.char <- ContTime$Z.char
Time.char <- ContTime$Time.char
fit <- Strat.cox(data, Event.char, Z.char, Time.char, prov.char, group = c(1, 2, 2, 3, 3))
predict(fit, data, Z.char, lambda = fit$lambda, type = "response")[1:5, 1:5]
}
