% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_profile_lkd.R
\name{prof_lkd.linear}
\alias{prof_lkd.linear}
\title{Fit the linear model using profile likelihood}
\usage{
prof_lkd.linear(data, Y.char, Z.char, prov.char, ...)
}
\arguments{
\item{data}{an \code{dataframe} or \code{list} object that contains the variables in the model.}

\item{Y.char}{name of the response variable in \code{data} as a character string.}

\item{Z.char}{names of covariates in \code{data} as vector of character strings.}

\item{prov.char}{name of provider IDs variable in \code{data} as a character string. If "prov.char" is not specified, all observations are are considered to be from the same provider.}

\item{...}{extra arguments to be passed to function.}
}
\value{
An object with S3 class \code{proflkd.linear}

\item{beta}{the fitted matrix of covariate coefficients.}

\item{gamma}{the fitted matrix of provider effects.}
}
\description{
Fitting a linear model for data with high-dimensional center effect through profile likelihood estimation
}
\examples{
data(linear_data)
data <- linear_data$data
Y.char <- linear_data$Y.char
prov.char <- linear_data$prov.char
Z.char <- linear_data$Z.char
fit <- prof_lkd.linear(data, Y.char, Z.char, prov.char)
fit$beta[1:5, ]
fit$gamma[1:5, ]

}
