% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Surv_Data}
\alias{Surv_Data}
\title{example data for `pp.DiscSurv`}
\format{
A list containing the following elements:
\describe{
  \item{time}{observation time.}
  \item{status}{event indicator.}
  \item{Prov.ID}{provider information.}
  \item{Z1, ..., Z5}{5 continuous covariates.}
}
}
\usage{
data(Surv_Data)
}
\description{
A simulated data set containing time, status, provider information and 5 covariates.
}
\keyword{datasets}
