% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_ppDiscSurv.R
\name{coef.ppDiscSurv}
\alias{coef.ppDiscSurv}
\title{Extract coefficients of a ppDiscSurv object}
\usage{
\method{coef}{ppDiscSurv}(fit, lambda, which = 1:length(fit$lambda), drop = TRUE, ...)
}
\arguments{
\item{fit}{a \code{ppDiscSurv} object.}

\item{lambda}{values of the regularization parameter lambda at which coefficients are requested. For values of lambda not in the sequence of fitted models, linear interpolation is used.}

\item{which}{indices of the penalty parameter lambda at which predictions are required. By default, all indices are returned. If lambda is specified, this will override which.}

\item{drop}{whether to keep coefficient names}

\item{...}{}
}
\description{
Return the model coefficients of a \code{ppDiscSurv} object
}
\examples{
data(Surv_Data)
data <- Surv_Data$data
Event.char <- Surv_Data$Event.char
prov.char <- Surv_Data$prov.char
Z.char <- Surv_Data$Z.char
Time.char <- Surv_Data$Time.char
fit <- pp.DiscSurv(data, Event.char, prov.char, Z.char, Time.char)
coef(fit, lambda = fit$lambda)$beta[, 1:10]
coef(fit, lambda = fit$lambda)$gamma[, 1:10]
}
