% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.ppLasso}
\alias{coef.ppLasso}
\alias{coef.gr_ppLasso}
\title{extract coefficients of a ppLasso or gr_ppLasso object}
\usage{
\method{coef}{ppLasso}(fit, lambda, which = 1:length(fit$lambda), drop = TRUE, ...)

\method{coef}{gr_ppLasso}(fit, lambda, which = 1:length(fit$lambda), drop = TRUE, ...)
}
\arguments{
\item{fit}{a \code{ppLasso} or \code{gr_ppLasso} object.}

\item{lambda}{values of the regularization parameter lambda at which coefficients are requested. For values of lambda not in the sequence of fitted models, linear interpolation is used.}

\item{which}{indices of the penalty parameter lambda at which predictions are required. By default, all indices are returned. If lambda is specified, this will override which.}

\item{drop}{whether to keep coefficient names}

\item{...}{}
}
\description{
return the model coefficients of a \code{ppLasso} or \code{gr_ppLasso} object
}
\examples{
#fit glm without grouped covariates
data(GLM_Data)
data <- GLM_Data$data
Y.char <- GLM_Data$Y.char
prov.char <- GLM_Data$prov.char
Z.char <- GLM_Data$Z.char
fit <- pp.lasso(data, Y.char, Z.char, prov.char)
coef(fit, lambda = fit$lambda)$beta[, 1:10]
coef(fit, lambda = fit$lambda)$gamma[1:10, 1:5]
#fit glm with grouped covariates
data(GLM_Data)
data <- GLM_Data$data
Y.char <- GLM_Data$Y.char
prov.char <- GLM_Data$prov.char
Z.char <- GLM_Data$Z.char
group <- GLM_Data$group
fit <- grp.lasso(data, Y.char, Z.char, prov.char, group = group)
coef(fit, lambda = fit$lambda)$beta[, 1:5]
coef(fit, lambda = fit$lambda)$gamma[1:10, 1:5]
}
