% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_strat_cox.R
\name{coef.strat_cox}
\alias{coef.strat_cox}
\title{Extract coefficients of a strat_cox object}
\usage{
\method{coef}{strat_cox}(fit, lambda, which = 1:length(fit$lambda), drop = TRUE, ...)
}
\arguments{
\item{fit}{a \code{strat_cox} object.}

\item{lambda}{values of the regularization parameter lambda at which coefficients are requested. For values of lambda not in the sequence of fitted models, linear interpolation is used.}

\item{which}{indices of the penalty parameter lambda at which predictions are required. By default, all indices are returned. If lambda is specified, this will override which.}

\item{drop}{whether to keep coefficient names}

\item{...}{}
}
\description{
Return the model coefficients of a \code{strat_cox} object
}
\examples{
data(ContTime)
data <- ContTime$data
Event.char <- ContTime$Event.char
prov.char <- ContTime$prov.char
Z.char <- ContTime$Z.char
Time.char <- ContTime$Time.char
fit <- Strat.cox(data, Event.char, Z.char, Time.char, prov.char, group = c(1, 2, 2, 3, 3))
coef(fit, lambda = fit$lambda)[, 1:5]
}
