% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.DiscSurv.R
\name{cv.DiscSurv}
\alias{cv.DiscSurv}
\title{Cross-validation for DiscSurv}
\usage{
cv.DiscSurv(
  data,
  Event.char,
  Z.char,
  Time.char,
  penalize.x = rep(1, length(Z.char)),
  ...,
  nfolds = 10,
  seed,
  fold,
  trace.cv = FALSE
)
}
\arguments{
\item{data}{an `dataframe` or `list` object that contains the variables in the model.}

\item{Event.char}{name of the event indicator in `data` as a character string.}

\item{Z.char}{names of covariates in `data` as vector of character strings.}

\item{Time.char}{name of the observation time in `data` as a character string.}

\item{penalize.x}{a vector indicates whether the corresponding covariate will be penalized, as in \code{pp.DiscSurv} function.}

\item{...}{extra arguments to be passed to function.}

\item{nfolds}{the number of cross-validation folds. Default is 10.}

\item{seed}{the seed of the random number generator in order to obtain reproducible results.}

\item{fold}{a vector that specifies the fold that observations belongs to. By default the observations are randomly assigned.}

\item{trace.cv}{\code{cv.DiscSurv} will provide user with the progress of cross validation if `trace.cv = TRUE`. Default is FALSE.}
}
\value{
An object with S3 class \code{cv.DiscSurv}.

\item{cve}{the error for each value of lambda, averaged across the cross-validation folds.}

\item{cvse}{the estimated standard error associated with each value of for cve.}

\item{lambda}{the sequence of regularization parameter values along which the cross-validation error was calculated.}

\item{fit}{the fitted \code{pp.DiscSurv} object for the whole data.}

\item{fold}{the fold assignments for cross-validation for each observation}

\item{min}{the index of lambda corresponding to lambda.min.}

\item{lambda.min}{the value of lambda with the minimum cross-validation error.}
}
\description{
Performs k-fold cross validation for penalized regression models over a grid of values of regularization parameter lambda.
}
\examples{
data(Surv_Data)
data <- Surv_Data$data
Event.char <- Surv_Data$Event.char
Z.char <- Surv_Data$Z.char
Time.char <- Surv_Data$Time.char
cv.fit <- cv.DiscSurv(data, Event.char, Z.char, Time.char, nfolds = 10, trace.cv = T)
cv.fit$cve
cv.fit$lambda.min

}
\references{
K. He, J. Kalbfleisch, Y. Li, and et al. (2013) Evaluating hospital readmission rates in dialysis facilities; adjusting for hospital effects.
\emph{Lifetime Data Analysis}, \strong{19}: 490-512.
\cr
}
