% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{GLM_Data}
\alias{GLM_Data}
\title{Example data for generalize lienar model}
\format{
A list containing the following elements:
\describe{
  \item{data}{example data. \code{Y} is the response variable; \code{Prov.ID} is the provider indicator; \code{Z1}, ..., \code{Z5} are 5 continuous covariates.}
  \item{Y.char}{variable name of the response variable.}
  \item{prov.char}{variable name of the provider indicator.}
  \item{Z.char}{variable names of covariates.}
  \item{group}{vector describing how the covariates are grouped.}

}
}
\usage{
data(GLM_Data)
}
\description{
A simulated data set containing response variable, provider information and 5 covariates.
}
\keyword{datasets}
