\name{grplasso-internal}
\alias{blockstand}
\alias{create.design}
\title{Internal grplasso functions}
\description{Internal grplasso functions}
\usage{
blockstand(x, ipen.which, inotpen.which)
create.design(m, formula, nonpen  = ~ 1, data, weights, subset,
              na.action, contrasts, env)
}
\details{These are not to be called by the user.}
\keyword{internal}

