\name{grplasso-package}
\alias{grplasso-package}
\docType{package}
\title{
grplasso
}
\description{
Fits generalized linear models or user specified models with group lasso
penalty.}
\details{
\tabular{ll}{
Package: \tab grplasso\cr
Type: \tab Package\cr
Version: \tab 0.1-2\cr
Date: \tab 2006-08-11\cr
Depends: \tab methods\cr
License: \tab GPL\cr
Built: \tab R 2.3.1; ; 2006-06-01 09:38:41; unix\cr
}

Please note that this is an *early test release*.

The best entry point for the package are the examples in the help
file of the function \code{\link{grplasso}}.

Index:
\preformatted{
grpl.control            Options for the Group Lasso Algorithm
grpl.control-class      Class "grpl.control": Options for the Group
                        Lasso Algorithm
grpl.model              Group Lasso models
grpl.model-class        Class "grpl.model": Group Lasso Models
grplasso                Function to fit a solution of a group lasso
                        problem
grplasso-package        grplasso
lambdamax               Function to find the maximal value of the
                        penalty parameter lambda
plot.grplasso           Plots the solution path of a grplasso object
predict.grplasso        Predict method for grplasso objects
splice                  Dataset of human donor splice sites
}
}
\author{
Lukas Meier

Maintainer: Lukas Meier <meier@stat.math.ethz.ch>
}
\keyword{package}
