% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cve.Discsurv.R
\name{plot.cv.DiscSurv}
\alias{plot.cv.DiscSurv}
\title{Plot the cross entropy loss from a cv.DiscSurv object}
\usage{
\method{plot}{cv.DiscSurv}(
  fit,
  log.x = T,
  vertical.line = T,
  col.vertical.line = "blue",
  col.dot = "red"
)
}
\arguments{
\item{fit}{a \code{cv.DiscSurv} object.}

\item{log.x}{whether the horizontal axis be on the log scale.}

\item{vertical.line}{whether draws a vertical line at the value where cross-validaton error is minimized.}
}
\description{
Return the plot of the cross entropy loss from a \code{cv.DiscSurv} object
}
\examples{
data(DiscTime)
data <- DiscTime$data
Event.char <- DiscTime$Event.char
Z.char <- DiscTime$Z.char
Time.char <- DiscTime$Time.char
cv.fit.DiscSurv <- cv.DiscSurv(data, Event.char, Z.char, Time.char, nfolds = 10)
plot(cv.fit.DiscSurv)
}
