% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cve.ppDiscsurv.R
\name{plot.cv.ppDiscSurv}
\alias{plot.cv.ppDiscSurv}
\title{plot the cross entropy Loss from a cv.ppDiscSurv object}
\usage{
\method{plot}{cv.ppDiscSurv}(
  fit,
  log.x = T,
  vertical.line = T,
  col.vertical.line = "blue",
  col.dot = "red"
)
}
\arguments{
\item{fit}{a \code{cv.ppDiscSurv} object.}

\item{log.x}{whether the horizontal axis be on the log scale.}

\item{vertical.line}{whether draws a vertical line at the value where cross-validaton error is minimized.}
}
\description{
return the plot of the cross entropy loss from a \code{cv.ppDiscSurv} object
}
\examples{
data(Surv_Data)
data <- Surv_Data$data
Event.char <- Surv_Data$Event.char
prov.char <- Surv_Data$prov.char
Z.char <- Surv_Data$Z.char
Time.char <- Surv_Data$Time.char
cv.fit.ppDiscSurv <- cv.pp.DiscSurv(data, Event.char, prov.char, Z.char, Time.char, nfolds = 10)
plot(cv.fit.ppDiscSurv)
}
