% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cve.ppLasso.R
\name{plot.cv.ppLasso}
\alias{plot.cv.ppLasso}
\alias{plot.cv.gr_ppLasso}
\title{plot the cross entropy Loss from a cv.ppLasso or cv.gr_ppLasso object}
\usage{
\method{plot}{cv.ppLasso}(
  fit,
  log.x = T,
  vertical.line = T,
  col.vertical.line = "blue",
  col.dot = "red"
)

\method{plot}{cv.gr_ppLasso}(
  fit,
  log.x = T,
  vertical.line = T,
  col.vertical.line = "blue",
  col.dot = "red"
)
}
\arguments{
\item{fit}{a \code{cv.gr_ppLasso} object.}

\item{log.x}{whether the horizontal axis be on the log scale.}

\item{vertical.line}{whether draws a vertical line at the value where cross-validaton error is minimized.}
}
\description{
return the plot of the cross entropy loss from a \code{cv.ppLasso} or \code{cv.gr_ppLasso} object
}
\examples{
data(GLM_Data)
data <- GLM_Data$data
Y.char <- GLM_Data$Y.char
prov.char <- GLM_Data$prov.char
Z.char <- GLM_Data$Z.char
cv.fit.pplasso <- cv.pp.lasso(data, Y.char, Z.char, prov.char, nfolds = 10)
plot(cv.fit.pplasso)
data(GLM_Data)
data <- GLM_Data$data
Y.char <- GLM_Data$Y.char
prov.char <- GLM_Data$prov.char
Z.char <- GLM_Data$Z.char
group <- GLM_Data$group
cv.fit.grplasso <- cv.grp.lasso(data, Y.char, Z.char, prov.char, group = group, nfolds = 10)
plot(cv.fit.grplasso)

}
