% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cve.strat_cox.R
\name{plot.cv.strat_cox}
\alias{plot.cv.strat_cox}
\title{Plot the cross entropy loss from a cv.strat_cox object}
\usage{
\method{plot}{cv.strat_cox}(
  fit,
  log.x = T,
  vertical.line = T,
  col.vertical.line = "blue",
  col.dot = "red"
)
}
\arguments{
\item{fit}{a \code{cv.strat_cox} object.}

\item{log.x}{whether the horizontal axis be on the log scale.}

\item{vertical.line}{whether draws a vertical line at the value where cross-validaton error is minimized.}
}
\description{
Return the plot of the cross entropy loss from a \code{cv.strat_cox} object
}
\examples{
data(Cox_Data)
data <- Cox_Data$data
Event.char <- Cox_Data$Event.char
prov.char <- Cox_Data$prov.char
Z.char <- Cox_Data$Z.char
Time.char <- Cox_Data$Time.char
cv.fit.strat_cox <- cv.strat_cox(data, Event.char, Z.char, Time.char, prov.char, group = c(1, 2, 2, 3, 3), nfolds = 10)
plot(cv.fit.strat_cox)
}
