% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.path.Discsurv.R
\name{plot.DiscSurv}
\alias{plot.DiscSurv}
\title{Plot regularization path of coefficients from a DiscSurv object}
\usage{
\method{plot}{DiscSurv}(fit, log.x = T, label = F)
}
\arguments{
\item{fit}{a \code{DiscSurv} object.}

\item{log.x}{whether the horizontal axis be on the log scale.}

\item{label}{whether annotates the plot with labels.}
}
\description{
Return the plot the regularization path from a \code{DiscSurv} object
}
\examples{
data(Surv_Data)
data <- Surv_Data$data
Event.char <- Surv_Data$Event.char
Z.char <- Surv_Data$Z.char
Time.char <- Surv_Data$Time.char
fit <- DiscSurv(data, Event.char, Z.char, Time.char)
plot(fit, label = T)
}
