% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.path.ppDiscsurv.R
\name{plot.ppDiscSurv}
\alias{plot.ppDiscSurv}
\title{plot regularization path of coefficients from a ppDiscSurv object}
\usage{
\method{plot}{ppDiscSurv}(fit, log.x = T, label = F)
}
\arguments{
\item{fit}{a \code{ppDiscSurv} object.}

\item{log.x}{whether the horizontal axis be on the log scale.}

\item{label}{whether annotates the plot with labels.}
}
\description{
return the plot the regularization path from a \code{ppDiscSurv} object
}
\examples{
data(Surv_Data)
data <- Surv_Data$data
Event.char <- Surv_Data$Event.char
prov.char <- Surv_Data$prov.char
Z.char <- Surv_Data$Z.char
Time.char <- Surv_Data$Time.char
fit <- pp.DiscSurv(data, Event.char, prov.char, Z.char, Time.char)
plot(fit, label = T)
}
