% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.path.ppLasso.R
\name{plot.ppLasso}
\alias{plot.ppLasso}
\alias{plot.gr_ppLasso}
\title{Plot regularization path of coefficients from a ppLasso or gr_ppLasso object}
\usage{
\method{plot}{ppLasso}(fit, log.x = T, label = F)

\method{plot}{gr_ppLasso}(fit, log.x = T, label = F)
}
\arguments{
\item{fit}{a \code{gr_ppLasso} object.}

\item{log.x}{whether the horizontal axis be on the log scale.}

\item{label}{whether annotates the plot with labels.}
}
\description{
Return the plot the regularization path from a \code{ppLasso} or \code{gr_ppLasso} object
}
\examples{
data(GLM_Data)
data <- GLM_Data$data
Y.char <- GLM_Data$Y.char
prov.char <- GLM_Data$prov.char
Z.char <- GLM_Data$Z.char
fit <- pp.lasso(data, Y.char, Z.char, prov.char)
plot(fit, label = T)
data(GLM_Data)
data <- GLM_Data$data
Y.char <- GLM_Data$Y.char
prov.char <- GLM_Data$prov.char
Z.char <- GLM_Data$Z.char
group <- GLM_Data$group
fit <- grp.lasso(data, Y.char, Z.char, prov.char, group = group)
plot(fit, label = T)
}
