% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.path.R
\name{plot.ppLasso}
\alias{plot.ppLasso}
\title{plot regularization path of coefficients from a `ppLasso` object}
\usage{
\method{plot}{ppLasso}(fit, log.x = T, label = F)
}
\arguments{
\item{fit}{a `ppLasso` object.}

\item{log.x}{whether the horizontal axis be on the log scale.}

\item{label}{whether annotates the plot with labels.}
}
\description{
plot the cross entropy Loss from a `ppLasso` object
}
\examples{
data(GLM_Data)
Y.char <- 'Y'
prov.char <- 'Prov.ID'
Z.char <- c("Z1", "Z2", "Z3", "Z4", "Z5")
fit <- pp.lasso(GLM_Data, Y.char, Z.char, prov.char)
plot(fit)
}
