% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.path.strat_cox.R
\name{plot.strat_cox}
\alias{plot.strat_cox}
\title{Plot regularization path of coefficients from a strat_cox object}
\usage{
\method{plot}{strat_cox}(fit, log.x = T, label = F)
}
\arguments{
\item{fit}{a \code{strat_cox} object.}

\item{log.x}{whether the horizontal axis be on the log scale.}

\item{label}{whether annotates the plot with labels.}
}
\description{
Return the plot the regularization path from a \code{strat_cox} object
}
\examples{
data(Cox_Data)
data <- Cox_Data$data
Event.char <- Cox_Data$Event.char
prov.char <- Cox_Data$prov.char
Z.char <- Cox_Data$Z.char
Time.char <- Cox_Data$Time.char
fit <- Strat.cox(data, Event.char, prov.char, Z.char, Time.char, group = c(1, 2, 2, 3, 3))
plot(fit, label = T)
}
