% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ppDiscSurv.R
\name{predict.ppDiscSurv}
\alias{predict.ppDiscSurv}
\title{predictions of a ppDiscSurv object}
\usage{
\method{predict}{ppDiscSurv}(
  fit,
  data,
  prov.char,
  Z.char,
  Time.char,
  lambda,
  which = 1:length(fit$lambda),
  type = c("response", "vars", "nvars"),
  return.Array = TRUE,
  ...
)
}
\arguments{
\item{fit}{a \code{ppDiscSurv} object.}

\item{data}{an `dataframe` or `list` object that contains the variables for prediction.}

\item{prov.char}{name of provider IDs variable in `data` as a character string.}

\item{Z.char}{names of covariates in `data` as vector of character strings.}

\item{Time.char}{name of the observation time in `data` as a character string.}

\item{lambda}{values of the regularization parameter lambda at which predictions are requested. For values of lambda not in the sequence of fitted models, linear interpolation is used.}

\item{which}{indices of the penalty parameter lambda at which predictions are required. By default, all indices are returned. If lambda is specified, this will override which.}

\item{type}{type of prediction: \code{response} provides the fitted value of each person at each time point ;
\code{vars} returns the indices for the non-zero coefficients; \code{nvars} returns the number of non-zero coefficients;}

\item{...}{}
}
\description{
return the model predictions of a \code{ppDiscSurv} object
}
\examples{
data(Surv_Data)
data <- Surv_Data$data
Event.char <- Surv_Data$Event.char
prov.char <- Surv_Data$prov.char
Z.char <- Surv_Data$Z.char
Time.char <- Surv_Data$Time.char
fit <- pp.DiscSurv(data, Event.char, prov.char, Z.char, Time.char)
predict(fit, data, Z.char, prov.char, lambda = fit$lambda, type = "response")[1:10, 1:5]
predict(fit, data, Z.char, prov.char, lambda = 0.04, type = "vars")
}
